refactor(iocuak): extract container module module
