+++ packages/iocuak/src/container/services/api/ContainerServiceApiImplementation.spec.ts
@@ -7,6 +7,7 @@ jest.mock('../../../metadata/utils/api/convertBindingToBindingApi');
 
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
+import { ContainerModuleApi } from '../../../containerModule/models/api/ContainerModuleApi';
 import { ContainerModuleMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleMetadataApi';
 import { ContainerModuleMetadata } from '../../../containerModuleTask/models/domain/ContainerModuleMetadata';
 import { convertToContainerModuleMetadata } from '../../../containerModuleTask/utils/api/convertToContainerModuleMetadata';
@@ -16,7 +17,6 @@ import { Binding } from '../../../metadata/models/domain/Binding';
 import { BindingType } from '../../../metadata/models/domain/BindingType';
 import { MetadataService } from '../../../metadata/services/domain/MetadataService';
 import { convertBindingToBindingApi } from '../../../metadata/utils/api/convertBindingToBindingApi';
-import { ContainerModuleApi } from '../../modules/api/ContainerModuleApi';
 import { bind } from '../../utils/bind';
 import { bindToValue } from '../../utils/bindToValue';
 import { ContainerBindingService } from '../domain/ContainerBindingService';
