+++ packages/iocuak/src/container/services/api/ContainerServiceApiImplementation.ts
@@ -1,12 +1,12 @@
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
+import { ContainerModuleApi } from '../../../containerModule/models/api/ContainerModuleApi';
 import { ContainerModuleMetadataApi } from '../../../containerModuleTask/models/api/ContainerModuleMetadataApi';
 import { ContainerModuleMetadata } from '../../../containerModuleTask/models/domain/ContainerModuleMetadata';
 import { convertToContainerModuleMetadata } from '../../../containerModuleTask/utils/api/convertToContainerModuleMetadata';
 import { BindingApi } from '../../../metadata/models/api/BindingApi';
 import { Binding } from '../../../metadata/models/domain/Binding';
 import { convertBindingToBindingApi } from '../../../metadata/utils/api/convertBindingToBindingApi';
-import { ContainerModuleApi } from '../../modules/api/ContainerModuleApi';
 import { bind } from '../../utils/bind';
 import { bindToValue } from '../../utils/bindToValue';
 import { ContainerService } from '../domain/ContainerService';
