+++ packages/iocuak/src/containerModule/models/api/ContainerModuleApi.ts
@@ -1,4 +1,4 @@
-import { ContainerModuleBindingServiceApi } from '../../services/api/ContainerModuleBindingServiceApi';
+import { ContainerModuleBindingServiceApi } from '../../../container/services/api/ContainerModuleBindingServiceApi';
 
 export interface ContainerModuleApi {
   load(containerModuleBindingService: ContainerModuleBindingServiceApi): void;
