+++ packages/iocuak/src/containerModule/utils/api/convertToContainerModule.spec.ts
@@ -3,13 +3,13 @@ jest.mock('../../../container/utils/bindToValue');
 
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
-import { ContainerModuleApi } from '../../../container/modules/api/ContainerModuleApi';
-import { ContainerModule } from '../../../container/modules/domain/ContainerModule';
 import { ContainerModuleBindingServiceApi } from '../../../container/services/api/ContainerModuleBindingServiceApi';
 import { ContainerBindingService } from '../../../container/services/domain/ContainerBindingService';
 import { bind } from '../../../container/utils/bind';
 import { bindToValue } from '../../../container/utils/bindToValue';
 import { MetadataService } from '../../../metadata/services/domain/MetadataService';
+import { ContainerModuleApi } from '../../models/api/ContainerModuleApi';
+import { ContainerModule } from '../../models/domain/ContainerModule';
 import { convertToContainerModule } from './convertToContainerModule';
 
 describe(convertToContainerModule.name, () => {
