+++ packages/iocuak/src/containerModule/utils/api/convertToContainerModule.ts
@@ -1,12 +1,12 @@
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
-import { ContainerModuleApi } from '../../../container/modules/api/ContainerModuleApi';
-import { ContainerModule } from '../../../container/modules/domain/ContainerModule';
 import { ContainerModuleBindingServiceApi } from '../../../container/services/api/ContainerModuleBindingServiceApi';
 import { ContainerBindingService } from '../../../container/services/domain/ContainerBindingService';
 import { bind } from '../../../container/utils/bind';
 import { bindToValue } from '../../../container/utils/bindToValue';
 import { MetadataService } from '../../../metadata/services/domain/MetadataService';
+import { ContainerModuleApi } from '../../models/api/ContainerModuleApi';
+import { ContainerModule } from '../../models/domain/ContainerModule';
 
 export function convertToContainerModule(
   containerModuleApi: ContainerModuleApi,
