+++ packages/iocuak/src/containerModuleTask/models/cuaktast/ContainerModuleLoadFromMetadataTask.ts
@@ -1,8 +1,8 @@
 import * as cuaktask from '@cuaklabs/cuaktask';
 
-import { ContainerModule } from '../../../container/modules/domain/ContainerModule';
 import { ContainerBindingService } from '../../../container/services/domain/ContainerBindingService';
 import { ContainerInstanceService } from '../../../container/services/domain/ContainerInstanceService';
+import { ContainerModule } from '../../../containerModule/models/domain/ContainerModule';
 import { MetadataService } from '../../../metadata/services/domain/MetadataService';
 import { ContainerModuleClassMetadata } from '../domain/ContainerModuleClassMetadata';
 import { ContainerModuleFactoryMetadata } from '../domain/ContainerModuleFactoryMetadata';
